moving the ClickAwayListener
